Switch to docsRouter
